Tests/LibWeb/Text/expected/css/scale-number-percentage-interpolation.txt

100% -> 200%: 1.5
1 -> 200%: 1.5
100% -> 2: 1.5
none -> 200%: 1.5
